beginning of the config file parsing
