+++ src/WebServ/WebServerProg.hpp
@@ -1,12 +1,30 @@
 
 
-#ifndef WebServerProg_HPP
-# define WebServerProg_HPP
+#ifndef WEBSERVERPROG_H
+#define WEBSERVERPROG_H
 
 # include <iostream>
 # include <vector>
 # include <map>
 
+struct location
+{
+    std::vector<int> allowedMethods;
+    std::string locationPath;
+    std::string redirection;
+    std::string root;
+    int listing;
+    std::string defaultFile;
+    std::string cgiPath;
+};
+
+struct server
+{
+    int client_max_body_size;
+    std::string serverName;
+    std::map<int, std::string> errorPages; //this might be over kill
+    std::vector<location> locations; 
+};
 
 
 class WebServerProg
@@ -14,8 +32,9 @@ class WebServerProg
 	private:
 		std::vector<struct pollfd> m_pollSocketsVec;// this is where the polling will happen
 		//parse the configfile
-		//vector<maps<string, string>> servers
+		std::vector<struct server> servers;
 		size_t serverCount;
+		const char *defaultFileName;
 		
 	public:
 
@@ -33,6 +52,7 @@ class WebServerProg
 
 };
 
-std::vector parseConfigFile(std<string> fileName);
+//std::vector<struct server>
+void parseConfigFile(const char* fileName);
 
 #endif
